pyproject.toml
[tool.poetry]
name = "vx-underground-processor"
version = "0.1.0"
description = "Utility for downloading and processing files from VX Underground"
authors = ["Artur Iasenovets <t.me/artias13>"]
readme = "README.md"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
requests = "^2.27"
pyyaml = "^6.0"
yara-python = "^4.2"
apache-airflow = "^2.10.3"
minio = "^7.2.10"
python-dotenv = "^1.0.1"
tqdm = "^4.67.0"
py7zr = "^0.22.0"
